pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "ninja", "scikit-build-core>=0.9"]
build-backend = "setuptools.build_meta"

[project]
name = "earli"
version = "0.1.0"
description = "EARLI: Evolutionary Algorithm with RL Initialization for VRP"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "numpy==2.2.6",
  "scipy==1.14.1",
  "pandas==2.3.3",
  "matplotlib==3.10.6",
  "seaborn==0.13.2",
  "opencv-python==4.12.0.88",
  "colorama==0.4.6",
  "shapely==2.1.2",
  "pyproj>=3.6",
  "geopandas==1.1.1",
  "rasterio>=1.3",
  "xyzservices>=2024.4",
  "mercantile>=1.2",
  "contextily==1.6.2",

  # Torch family (from your env.yaml)
  "torch==2.8.0",
  "torchvision==0.23.0",
  "tensordict==0.10.0",
  "torchrl==0.10.0",
  "einops==0.8.1",

  "gymnasium==1.2.1",
  "stable-baselines3==2.7.0",

  "tqdm==4.67.1",
  "requests==2.32.5",
  "ruamel-yaml==0.18.15",
  "pyyaml==6.0.3",
  "wandb==0.22.1",

  # cuOpt SDK — align to env.yaml (25.8.*) and let it pull RAPIDS/CUDA bits
  "cuopt-cu12==25.8.*",

  # build helper + your HGS repo
  "ninja>=1.11",
  "hygese @ git+https://gitlab.com/igreenberg/initializable_hgs.git",
]

[project.urls]
homepage = "https://github.com/NVlabs/EARLI"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["earli*"]

[tool.setuptools.package-data]
earli = ["pretrained_models/*"]